This is an initial checkin of an Optionset model with matching manager.
Currently a standalone server, it is to be incorporated into the rest
of AdHoc soon, therefore the "tmp" folder.

The Optionset model has dynamic attributes, which the OptionsetManager
reads from the database on server startup. All templates, updates and
searches are added dynamically at the same time (using RPCC's new
@auto_template/@auto_update/@auto_search feature).

Still to do is to implement the methods to create, change and remove
options in the latest API version if the server is at "development"
deploy level (detecting this reliably is also to be included), and
methods to list all options for a particular API version.

Note: The client expects there to be two optionsets with ID:s 1 and
2. It updates, fetches and digs them to test the correct workings of
the server part.



git-svn-id: https://svn.ita.chalmers.se/repos/adhoc/trunk@184 9cf5bdc2-3f39-413b-b911-f617d3cba2d0