dconf:
Fixed syntax error in dconf chhost.
Changed help text for chhost to point out that a wildcard is available when changing hosts.
Added a -q flag to chhost to be able to disable questiuon when using -a flag or wildcard.

AdHoc server:
Changed requirement for host update info to make it possible to set info to null.

git-svn-id: https://svn.ita.chalmers.se/repos/adhoc/trunk@748 9cf5bdc2-3f39-413b-b911-f617d3cba2d0