Here's the commit message for the previous commit:
Added possibility to request the current members for a group. This works recursively sownwards the group tree.

Fixed two bugs in rpcc/ExtTypes.py that caised internal errors of the wrong kind when output fromn the server did not pass the checks.

This commit:

Lost an import in host.py that was moved.

git-svn-id: https://svn.ita.chalmers.se/repos/adhoc/trunk@777 9cf5bdc2-3f39-413b-b911-f617d3cba2d0